requirements.txt
tensorflow~=2.16.1
pillow~=10.3.0
keras~=3.3.3
pillow~=10.3.0
scipy~=1.13.0
scikit-learn~=1.4.2
scikit-learn-intelex~=2024.4.0
pandas~=2.2.2
numpy~=1.26.4
mlxtend~=0.23.1
pillow~=10.3.0

jupyter~=1.0.0
jupyter-console~=6.6.3
jupyter-events~=0.10.0
jupyter-lsp~=2.2.5
jupyter-nbextensions-configurator~=0.6.3
jupyter-ydoc~=0.3.4
jupyter_client~=8.6.1
jupyter_contrib_core~=0.4.2
jupyter_core~=5.7.2
jupyter_server~=2.14.0
jupyter_server_fileid~=0.9.2
jupyter_server_terminals~=0.5.3
jupyter_server_ydoc~=0.8.0
jupyterlab~=4.1.8
jupyterlab_pygments~=0.3.0
jupyterlab_server~=2.27.1
jupyterlab_widgets~=3.0.10

ipycanvas~=0.13.2
ipykernel~=6.29.4
ipympl~=0.9.4
ipython~=8.24.0
ipywidgets~=8.1.2
